rename objective actions file
